Moved to passing the entire Event instead of just the Key

+++ Makefile
@@ -6,4 +6,4 @@ all: game
 game:
 	CGO_ENABLED=0 GOOS=linux go build -o ./$(NAME) ./cmd/
 clean:
-	rm $(NAME)
+	@[ -f $(NAME) ] && rm $(NAME) || true
